fix(cuda): widen record and source time-series indices to 64-bit
`(b * n + i) * solver.nt + it` was computed in int32. With a full-survey
receiver set (106,615 cells) and the long window a 615-bin comb needs
(nt = 47,397), the product reaches 5.05e9 and wraps negative, so the kernel
writes outside the record buffer: "CUDA error: an illegal memory access was
encountered" on the first forward step.

Domain decomposition hid this for a long time -- each rank only owns its
tile's receivers, which kept the per-rank product under the limit. A
single-GPU run of the same configuration is what exposed it.

Same class as the boundary-saver overflow already fixed in
bug/boundary-kernel-int-idx-overflow, in the record and source kernels
instead.

The multiplications are promoted before they can wrap, not cast afterwards.
`spatial_size` / `u_idx` are widened alongside in the same kernel family so
the index width does not vary within it.

Bit-exact where nothing overflowed: record, gradient and loss hashes are
identical to dev on a 3-D acoustic forward+backward, and the criterion was
shown attainable first (same tree twice gives the same hashes). Full suite
run sequentially on both trees: 712 passed / 1 failed on each, the same
pre-existing test_datasets_registry failure.

+++ src/sweep/csrc/cuda/equations/elastic2d/backward.cu
@@ -50,10 +50,10 @@ static __global__ void sub_receiver_rho_grad_correction_apm_2d(
         iz < halo || iz >= solver.nz - halo)
         return;
 
-    int spatial_size = solver.nx * solver.nz;
+    long long spatial_size = (long long)solver.nx * solver.nz;
     int idx = iz * solver.nx + ix;
-    int u_idx = b * spatial_size + idx;
-    int src_idx = (b * nrec + s) * solver.nt + it;
+    long long u_idx = (long long)b * spatial_size + idx;
+    long long src_idx = ((long long)b * nrec + s) * solver.nt + it;
 
     float drho_x_drho, drho_z_drho;
     apm_rho_jacobian(category[idx], &drho_x_drho, &drho_z_drho);
